+++ src/pages/administration/students/index.vue
@@ -337,8 +337,11 @@ export default {
 				);
 			}
 
-			// filters out the consent column if ADMIN_TABLES_DISPLAY_CONSENT_COLUMN env is disabled
-			if (!process.env["ADMIN_TABLES_DISPLAY_CONSENT_COLUMN"]) {
+			// filters out the consent column if ADMIN_TABLES_DISPLAY_CONSENT_COLUMN env is disabled and school if not external
+			if (
+				!process.env["ADMIN_TABLES_DISPLAY_CONSENT_COLUMN"] &&
+				this.schoolInternallyManaged
+			) {
 				editedColumns = editedColumns.filter(
 					(col) => col.field !== "consentStatus"
 				);