+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Added
 
-- SC-8681 - added ADMIN_TABLES_DISPLAY_CONSENT_COLUMN as an environment variable to display/hide the NAT consent column
+- SC-8681 - added ADMIN_TABLES_DISPLAY_CONSENT_COLUMN as an environment variable to display/hide the NAT consent column. If it's disabled and the school is not externally managed, the column will be hidden.
 
 ### Fixed